app_iphone/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
apscheduler==3.10.4
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic-settings==2.6.1
pydantic==2.10.3
email-validator==2.2.0
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
anyio==4.7.0
